@@ -1,16 +1,17 @@
 buildscript {
-    ext.kotlin_version = '1.2.51'
+    ext.kotlin_version = '1.3.41'
 
     dependencies {
-        classpath 'org.owasp:dependency-check-gradle:5.0.0-M2'
+        classpath 'org.owasp:dependency-check-gradle:5.2.0'
     }
 }
 
 plugins {
-    id "org.jetbrains.kotlin.plugin.spring" version "1.2.51"
-    id 'org.jetbrains.kotlin.jvm' version "1.2.51"
-    id 'com.github.johnrengelman.shadow' version '4.0.3'
-    id 'io.spring.dependency-management' version '1.0.6.RELEASE'
+    id "org.jetbrains.kotlin.plugin.spring" version "1.3.41"
+    id 'org.jetbrains.kotlin.jvm' version "1.3.41"
+    id 'com.github.johnrengelman.shadow' version '5.1.0'
+    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
+    id 'org.springframework.boot' version '2.1.7.RELEASE'
 }
 
 apply plugin: 'org.owasp.dependencycheck'
@@ -29,6 +30,13 @@ version = gradle.ext.version
 
 shadowJar {
     mergeServiceFiles('META-INF/spring.*')
+}
+
+test {
+    useJUnitPlatform()
+    testLogging {
+        events "passed", "skipped", "failed"
+    }
 }
 
 tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
