@@ -8,18 +8,18 @@ import lt.boldadmin.nexus.api.type.valueobject.Coordinates
 import lt.boldadmin.nexus.api.type.valueobject.Location
 import lt.boldadmin.nexus.backend.handler.ProjectHandler
 import lt.boldadmin.nexus.backend.route.Routes
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.Mock
 import org.mockito.Mockito.lenient
-import org.mockito.junit.MockitoJUnitRunner
+import org.mockito.junit.jupiter.MockitoExtension
 import org.springframework.test.web.reactive.server.WebTestClient
 import reactor.core.publisher.Mono
 import reactor.core.publisher.toMono
-import kotlin.test.assertEquals
 
-@RunWith(MockitoJUnitRunner::class)
+@ExtendWith(MockitoExtension::class)
 class ProjectHandlerTest {
 
     @Mock
@@ -27,7 +27,7 @@ class ProjectHandlerTest {
 
     private lateinit var webClient: WebTestClient
 
-    @Before
+    @BeforeEach
     fun `Set up`() {
         val contextStub = create()
         lenient()
