@@ -5,16 +5,16 @@ import lt.boldadmin.nexus.api.service.worklog.duration.WorklogDurationService
 import lt.boldadmin.nexus.backend.handler.worklog.duration.WorklogDurationHandler
 import lt.boldadmin.nexus.backend.route.Routes
 import lt.boldadmin.nexus.backend.test.unit.handler.create
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.Mock
 import org.mockito.Mockito.lenient
-import org.mockito.junit.MockitoJUnitRunner
+import org.mockito.junit.jupiter.MockitoExtension
 import org.springframework.test.web.reactive.server.WebTestClient
-import kotlin.test.assertEquals
 
-@RunWith(MockitoJUnitRunner::class)
+@ExtendWith(MockitoExtension::class)
 class WorklogDurationHandlerTest {
 
     @Mock
@@ -22,7 +22,7 @@ class WorklogDurationHandlerTest {
 
     private lateinit var webClient: WebTestClient
 
-    @Before
+    @BeforeEach
     fun setUp() {
         val contextStub = create()
         lenient()
@@ -46,7 +46,7 @@ class WorklogDurationHandlerTest {
             .expectBody(Long::class.java)
             .returnResult()
 
-        assertEquals(duration, response.responseBody)
+        assertEquals(duration, response.responseBody!!)
     }
 
     @Test
@@ -63,7 +63,7 @@ class WorklogDurationHandlerTest {
             .expectBody(Long::class.java)
             .returnResult()
 
-        assertEquals(durationsSum, response.responseBody)
+        assertEquals(durationsSum, response.responseBody!!)
     }
 
     @Test
@@ -80,6 +80,6 @@ class WorklogDurationHandlerTest {
             .expectBody(Long::class.java)
             .returnResult()
 
-        assertEquals(durationsSum, response.responseBody)
+        assertEquals(durationsSum, response.responseBody!!)
     }
 }
