@@ -8,10 +8,10 @@ import org.springframework.context.support.GenericApplicationContext
 import org.springframework.http.server.reactive.ReactorHttpHandlerAdapter
 import org.springframework.web.cors.reactive.CorsWebFilter
 import org.springframework.web.server.adapter.WebHttpHandlerBuilder
-import reactor.ipc.netty.http.server.HttpServer
+import reactor.netty.http.server.HttpServer
 
 
-fun main(args: Array<String>) {
+fun main() {
 
     val context = GenericApplicationContext()
     initializeBeans(context)
@@ -19,7 +19,9 @@ fun main(args: Array<String>) {
 
     val httpHandler = getWebHttpHandler(context)
 
-    HttpServer.create(8070).startAndAwait(ReactorHttpHandlerAdapter(httpHandler))
+    HttpServer.create().port(8070).handle(ReactorHttpHandlerAdapter(httpHandler)).bindNow().apply {
+        onDispose().block()
+    }
 }
 
 private fun initializeBeans(context: GenericApplicationContext) {
