@@ -15,7 +15,7 @@ import javax.validation.Validation
 fun beans() = beans {
 
     bean("mongoTemplate") {
-        MongoTemplate(SimpleMongoDbFactory(ref(), ref<Environment>()["MONGO_DATABASE"])).apply {
+        MongoTemplate(SimpleMongoDbFactory(ref(), ref<Environment>()["MONGO_DATABASE"]!!)).apply {
             setWriteConcern(WriteConcern.ACKNOWLEDGED)
         }
     }
@@ -25,8 +25,8 @@ fun beans() = beans {
         MongoClient(
             ServerAddress(environment["MONGO_HOST"]),
             MongoCredential.createCredential(
-                environment["MONGO_USERNAME"], environment["MONGO_AUTH_DATABASE"],
-                environment["MONGO_PASSWORD"].toCharArray()
+                environment["MONGO_USERNAME"]!!, environment["MONGO_AUTH_DATABASE"]!!,
+                environment["MONGO_PASSWORD"]!!.toCharArray()
             ),
             MongoClientOptions.builder().build()
         )
